backend/requirements.txt
asgiref==3.11.1
dj-rest-auth==7.0.2
Django==5.2.11
django-allauth==65.14.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-dotenv==1.2.1
sqlparse==0.5.5
tzdata==2025.3
